Use a float for speed.

+++ app/routes/lines.rb
@@ -28,7 +28,7 @@ module Transitmix
       params do
         requires :name, type: String
         requires :coordinates, type: Array
-        optional :speed, type: Integer
+        optional :speed, type: Float
         optional :color, type: String
         optional :service_windows, type: Array
       end
